fix: app root detection system (#1460)
This PR fixes an issue with app root detection system treating symlinks
to files ad directories.

To avoid similar problems in the future we not only change the guard
check from `!....isFilie()` to .`...isDirectory()` but also encapsulate
the error in try catch block.

### How Has This Been Tested: 

check is test apps are still working and run this version of radon on
the project where the issue happened.

### How Has This Change Been Documented:

this is an internal change