fix: build stale errors when switching projects (#1520)
Fixes #1515

The issue was caused by the file watcher callback from the old
`DeviceSession` reading the new `LaunchConfiguration` with the new app
root, causing the fingerprints to mismatch.
This resulted in a state change for the device session, which caused a
state change for the (now disposed) previous `DeviceSessionsManager`,
which triggered `Project`'s state change via the delegate method.
This PR fixes the issue by preventing a disposed `DeviceSessionsManager`
from triggering state changes.

### How Has This Been Tested: 
- open a monorepo in Radon
- wait for the app to load
- switch approots
- the "Native dependencies changed" shouldn't trigger

### How Has This Change Been Documented:
bugfix