Fix Expo 54 support by correcting the resolve paths for @expo/cli module (#1522)
After updating to expo 54 we noticed that the @expo/cli package no
longer installs under `node_modules` but in the relative
`node_modules/expo/node_modules` location. This change must've been
introduced in between the latest Expo beta and the now public release.

While it isn't entirely clear why this is happening, `@expo/cli` has
never been a direct app dependency anyway and has always be indirectly
installed via `expo` package.

This PR makes it so that when resolving `@expo/cli` package we always
provide `expo` package location as the path. This results in the same
behavior as if the `expo` package was importing `@expo/cli` and allows
us to properly resolve the location of that package.

The "old" way of installing works as well, because node dependency
resolution always perform lookup in parent directories too. This makes
it possible for the `expo` package to import `@expo/cli` regardless
whether the latter is installed under its isolated `node_modules` folder
or somewhere up the directory hierarchy.

### How Has This Been Tested: 
1) create new expo-54 project (expo go) and run it – see it loads up
correctly
2) try prebuilding the project and running again
3) test expo 52 and expo 53 projects.