Use owner stacks for inspector functionality instead of overriding react renderer (#1660)
Since the removal of JSX transform's `__source` metadata, we were
relying on modification to react renderer that'd bring that metadata
back to modern versions of React. This approach required additional work
with every RN release and was quite brittle (i.e. we couldn't support
the cases where users would use a different version of React).

In recent versions the concept of owner stack got introduced which
allows us to access the needed location of JSX tags in runtime without
the metadata passed from the transform. In some early versions of the
owner stack work, we'd only be getting the pointer to the sources where
the component is defined and not where it is actually used which wasn't
ideal and made us give up on this idea.

We recently tried owner stack based approach and since its "official"
release in 19.1 it is stable and works ok for our usecases.

This PR adapts the owner-stack based logic for the inspector
functionality for React Native 0.80 and above.

Highlights of this PR:
1) in wrapper, we detect whether `_debugStack` field is available. If it
is, we extract the component stack based on it and pass that back as the
inspector request callback.
2) since the new debug stack is calculated in runtime, it points to the
bundled source. We add additional logic that detects the case when stack
points to the bundled file. When this happens, we use the debugger to
perform the source map lookup to give us the correct source of the
element.
3) the source map lookup uses a custom request
`RNIDE_findOriginalPosition` that is implemented by our debug session
proxy (both for the vscode-js-debug based implementation and the custom
CDP based implementation). Under the hood, it uses
`getPreferredUILocation` custom request that is exposed by
`vscode-js-debug`. We use a different name to make it easier to follow
the custom request call path and be explicit that the request with
`RNIDE` is handled within out code.
4) we update the wrapper logic to use `line0Based` and `column0Based`
naming for line and column fields to make it clear when we handle
different components of the system that use different numbering scheme
for line and column numbers. For example, `_debugStack` uses 1-based
numbering, but the source maps from `vscode-js-debug` use 0-based
numbering.
5) we delete the overwrites in `babel_transform` for the render for
versions 0.80 and above. We also comment out all the JSX-transformer
fixes in case metro starts for version 0.80+ since we no longer rely on
the transform-backed mechanism there
6) in order to skip JSX transformer fixes in specific versions, we add a
new environment variable to metro `RADON_IDE_RN_VERSION` that indicates
the version of React Native.
6) we delete all the custom renderer code for RN versions that no longer
need it
1) Run this on react native 0.81, 0.80 and expo-54 apps – check the
inspector (right click)
2) Run this on older version (i.e. 0.78) and also check the inspector to
verify it works there
3) I also checked 0.81 with old architecture to make sure it doesn't
impact the debug owner stack capabilities.
4) finally, tested this on RN 82 app with "customJSDebug" option enabled
in order to verify the new custom command works via `CDPDebugAdapter`
too.